Refactor plugin to provide "select page" functionality
This commit replaces the "create page" functionality with a "select page" feature that enables the user to choose an existing page in Figma for versioning. Additionally, a new "select page" view was added to the UI, and changes have been made to several other files to incorporate the new feature and improve usability. Finally, the package version was incremented to reflect these changes.